fix(AddressEdit): should expose setAreaCode method
